[REFACTOR] Modularizing Functions Unrelated to Tracking
